halfway fixed testing
