+++ test/checks1.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const fsp = require('fs').promises;
 const os = require('os');
 const util = require('util');
 
@@ -184,351 +183,6 @@ function checkUnknownConfigItems(option) {
   // add check for specific bad option
 }
 
-//
-//
-//
-//
-//
-//
-//
-//
-//
-//
-class Checkers {
-  constructor(options = {}) {
-    this.checkers = {
-      // header: {validator: checkHeader},
-      // patch: {validator: checkPatch, data: []},
-      // route: {validator: checkMetrics},
-      // gc: {validator: checkGarbageCollection},
-      // eventloop: {validator: checkEventloop},
-      // proc: {validator: checkProc},
-    };
-    const {
-      initialCheckers = [],
-      filename = 'route-metrics.log',
-    } = options;
-
-    this.filename = filename;
-
-    for (const {CheckerClass, constructorArgs} of initialCheckers) {
-      const checker = new CheckerClass(...constructorArgs);
-      this.add(checker);
-    }
-  }
-
-  add(checker) {
-    this.checkers[checker.type] = checker;
-  }
-
-  check(entries) {
-    for (let i = 0; i < entries.length; i++) {
-      const {type, entry} = entries[i];
-
-      if (this.checkers[type]) {
-        // passing the index allows the checker to handle ordering constraints
-        this.checkers[type].check(entry, {index: i});
-      }
-    }
-  }
-
-  getCountOfRequiredEntries() {
-    let numberOfLinesNeeded = 0;
-    const specificCountsNeeded = Object.create(null);
-    for (const key in this.checkers) {
-      const countNeeded = this.checkers[key].getCountOfRequiredEntries();
-      specificCountsNeeded[key] = countNeeded;
-      numberOfLinesNeeded += countNeeded;
-    }
-
-    return {numberOfLinesNeeded, specificCountsNeeded};
-  }
-
-  async waitForLogLines(options = {}) {
-    const {numberOfLinesNeeded, specificCountsNeeded} = this.getCountOfRequiredEntries();
-
-    const maxWaitTime = ('maxWaitTime' in options) ? options.maxWaitTime : 100;
-    const loopWaitTime = ('loopWaitTime' in options) ? options.loopWaitTime : 10;
-    const maxTries = Math.ceil(maxWaitTime / loopWaitTime);
-    const start = Date.now();
-
-    let lines;
-    let text;
-    let foundAll = false;
-    for (let tries = 0; tries < maxTries; tries++) {
-      text = await fsp.readFile(this.filename, {encoding: 'utf8'});
-      lines = text.split('\n').slice(0, -1);
-      const logObjects = lines.map(l => JSON.parse(l));
-      if (lines.length >= numberOfLinesNeeded && this.haveAllRequestedLines(logObjects, specificCountsNeeded)) {
-        foundAll = true;
-        break;
-      } else {
-        await this.wait(loopWaitTime);
-      }
-    }
-    return {
-      text,
-      lines,
-      numberOfLinesNeeded,
-      foundAll,
-      waitTime: Date.now() - start
-    };
-  }
-
-  /**
- * allRequestLines() returns true if the minimum number of each required
- * type of log entry has been found. because it can be called multiple
- * times, but is called with all log entries - not just the new entries -
- * it does not modify the `needed` object; it copies it each time.
- */
-  haveAllRequestedLines(logObjects, specificCountsNeeded) {
-    const defaultsNeeded = {
-      header: 1,
-      patch: 0,
-      gc: 0,
-      eventloop: 0,
-      metric: 0,
-    };
-    // copy
-    const neededCounts = Object.assign({}, defaultsNeeded, specificCountsNeeded);
-    // get rid of keys that are not needed
-    for (const key in neededCounts) {
-      if (neededCounts[key] === 0) {
-        delete neededCounts[key];
-      }
-    }
-
-    // loop through each line of the file (it's been JSON.parsed)
-    for (const line of logObjects) {
-      if (line.type in neededCounts) {
-        neededCounts[line.type] -= 1;
-        // if the needCount is 0 then no more are needed.
-        if (neededCounts[line.type] <= 0) {
-          delete neededCounts[line.type];
-          // all needs met?
-          if (Object.keys(neededCounts).length === 0) {
-            return true;
-          }
-        }
-      }
-    }
-
-    return false;
-  }
-
-  wait(ms) {
-    return new Promise((resolve) => {
-      setTimeout(resolve, ms);
-    });
-  }
-}
-
-class HeaderChecker {
-  // makeLogEntryChecker('header', pdj, overrides)
-  constructor(pdj, overrides) {
-    this.type = 'header';
-    this.pdj = pdj;
-    this.overrides = overrides;
-  }
-
-  check(entry) {
-    const expected = Object.assign({}, _expected, {os: _os}, {package_json: this.pdj});
-    // set some things that change quickly to just match
-    const {freemem, loadavg, uptime} = entry.os;
-    expect(freemem).a('number');
-    expected.os.freemem = freemem;
-    expect(loadavg).an('array');
-    expected.os.loadavg = loadavg;
-    expect(uptime).a('number');
-    expected.os.uptime = uptime;
-
-    const cpus = os.cpus();
-    expected.os.cpus = cpus.length;
-    expected.os.cpuModel = cpus[0].model;
-
-    const ni = os.networkInterfaces();
-    expected.os.networkInterfaces = Object.create(null);
-    for (const iface in ni) {
-      // extract minimal information
-      expected.os.networkInterfaces[iface] = ni[iface].map((ni) => {
-        return {
-          address: ni.address,
-          netmask: ni.netmask
-        };
-      });
-    }
-
-    expected.config = {
-      LOG_FILE: 'route-metrics.log',
-      OUTPUT_CONFIG: '',
-      EVENTLOOP: false,
-      EVENTLOOP_RESOLUTION: 20,
-      GARBAGE_COLLECTION: false,
-    };
-
-    // always set these because a test might change them.
-    expected.argv = process.argv;
-    expected.execArgv = process.execArgv;
-
-    if (this.overrides) {
-      let overrides = this.overrides;
-      if (typeof overrides === 'function') {
-        overrides = this.overrides();
-      }
-
-      for (const [key, obj] of Object.entries(overrides)) {
-        if (!(key in expected)) {
-          throw new Error(`override specifies invalid key "${key}"`);
-        }
-        if (typeof expected[key] === 'object') {
-          if (!Array.isArray(expected[key])) {
-            Object.assign(expected[key], obj);
-          } else {
-            expected[key] = obj;
-          }
-        } else {
-          expected[key] = obj;
-        }
-      }
-    }
-
-    expect(entry).eql(expected);
-  }
-
-  getCountOfRequiredEntries() {
-    return 1;
-  }
-}
-
-class PatchChecker {
-  constructor(options) {
-    this.type = 'patch';
-    const {requiredPatches = [], ...otherOptions} = options;
-    this.staticRequiredPatches = requiredPatches.slice();
-    this.requiredPatches = requiredPatches.slice();
-
-    const {
-      nonRequiredPatchesAllowed = false,
-      duplicatesAllowed = false
-    } = otherOptions;
-    this.nonRequiredPatchesAllowed = nonRequiredPatchesAllowed;
-    this.duplicatesAllowed = duplicatesAllowed;
-  }
-
-  check(entry) {
-    const {name} = entry;
-    if (!this.requiredPatches.length) {
-      return;
-    }
-
-    if (this.requiredPatches.includes(name)) {
-      // remove the patch from the list - option to allow duplicates if needed
-      this.requiredPatches.splice(this.requiredPatches.indexOf(name), 1);
-      return;
-    }
-
-    // it's not a required patch. are non-required patches allowed?
-    if (this.nonRequiredPatchesAllowed) {
-      return;
-    }
-
-    // no, they're not allowed.
-    expect(this.staticRequiredPatches).includes(name);
-  }
-
-  getCountOfRequiredEntries() {
-    return this.staticRequiredPatches.length;
-  }
-
-  static getMinimalPatchEntries(t) {
-    // it's impossible to have a server without http; even https requires it.
-    const patchNames = [];
-
-    // the expected sequence of patches is specific to each combination of agent and
-    // express. N.B. as of node 16.19.1, the agent cannot be detected by route-metrics,
-    // because it is being loaded via the '-r' or '--request' command line option. so
-    // it is being detected by looking at process.execArgv when the first module in the
-    // @contrast scope is loaded.
-    if (t.agentPresent === '@contrast/agent') {
-      patchNames.push('http');
-      //
-      // http and https (since the agent started using axios) are always required
-      // by the node agent even if the server doesn't use https.
-      //
-      // for some reason we do not see get a chance to patch https when the agent
-      // is loaded by node v18.19.0 up to 20.0.0. i think this is because the
-      // tests use --import to load route-metrics and the agent and that can load
-      // in a background thread. there is no fix for this short of updating
-      // route-metrics with esm support.
-      //
-      // perversely, even if https is not specified, it gets loaded by axios
-      // which @contrast/agent uses, so it's always present when the agent is.
-      if (t.loadProtos.includes('https')) {
-        // if the server is going to load https it will always be present
-        patchNames.push('https');
-      } else {
-        // the server doesn't load, only the agent. so the weird node-version
-        // exclusion is needed.
-        if (nodeIsGreaterThanV20 || nodeIsLessThanV18dot19) {
-          patchNames.push('https');
-        }
-      }
-      // the agent should be present too.
-      patchNames.push(t.agentPresent);
-    } else if (!t.agentPresent) {
-      //
-      // no agent, so http and/or https are the only patch entries.
-      //
-      // both the node-agent and express require http, so it will be present
-      const expressPresent = t.server === 'express';
-      if (t.loadProtos.includes('http') || expressPresent) {
-        patchNames.push('http');
-      }
-      if (t.loadProtos.includes('https')) {
-        patchNames.push('https');
-      }
-    } else {
-      // this used to support rasp-v3, but that's been removed.
-      throw new Error(`unexpected agentPresent value: ${t.agentPresent}`);
-    }
-
-    return patchNames;
-  }
-}
-
-
-class RouteChecker {
-  constructor(options = {}) {
-    this.type = 'route';
-    this.routesExpected = options.routesExpected || 0;
-    this.routeEntryCount = 0;
-  }
-
-  getCountOfRequiredEntries() {
-    return this.routesExpected;
-  }
-
-  check(entry) {
-    // we could check the entry but we won't get here unless it's a route, so
-    // until we add specific route-matching, there's no need.
-    this.routeEntryCount += 1;
-  }
-
-  getRouteEntryCount() {
-    return this.routeEntryCount;
-  }
-}
-
-//
-//
-//
-//
-//
-//
-//
-//
-//
-
 /**
  * The loosely defined checking functions.
  */
@@ -708,9 +362,4 @@ module.exports = {
   makePatchEntryCheckers,
   makeMetricsChecker,
   makeTimeSeriesCheckers,
-  // checker classes
-  Checkers,
-  HeaderChecker,
-  PatchChecker,
-  RouteChecker,
 };
